requirements.txt
numpy==1.20.1
pandas==1.3.0
Pillow==8.4.0
streamlit==1.1.0
